rename: 대소문자만 바뀐 파일명이 git에 인식되지 않아서 다른 이름 적용 (#101)
