+++ src/components/Icon/iconEmployee2.js
