+++ src/components/Icon/iconHome2.js
