+++ src/components/Icon/iconNotice2.js
