+++ src/components/Icon/iconProfile2.js
