+++ src/components/Icon/index.js
@@ -1,7 +1,7 @@
-import IconHome from './IconHome';
-import IconProfile from './IconProfile';
-import IconTimetable from './IconTimetable';
-import IconNotice from './IconNotice';
-import IconEmployee from './IconEmployee';
+import IconHome from './iconHome2';
+import IconProfile from './iconProfile2';
+import IconTimetable from './iconTimetable2';
+import IconNotice from './iconNotice2';
+import IconEmployee from './iconEmployee2';
 
 export { IconHome, IconProfile, IconTimetable, IconNotice, IconEmployee };
